communication: move hiddevice to separate package
So the u2fhid package does not depend on karalabe/hid, which is
optional and only useful on Desktop, not mobile. `gomobile` can fail
because of this dep.
