+++ cmd/bluetooth_upgrade/main.go
@@ -24,6 +24,7 @@ import (
 	"github.com/BitBoxSwiss/bitbox02-api-go/api/firmware"
 	"github.com/BitBoxSwiss/bitbox02-api-go/api/firmware/mocks"
 	"github.com/BitBoxSwiss/bitbox02-api-go/communication/u2fhid"
+	"github.com/BitBoxSwiss/bitbox02-api-go/communication/u2fhid/hiddevice"
 	"github.com/karalabe/hid"
 )
 
@@ -80,7 +81,7 @@ func main() {
 	hidDevice, err := deviceInfo.Open()
 	errpanic(err)
 
-	comm := u2fhid.NewCommunication(u2fhid.NewHidDevice(hidDevice), bitboxCMD)
+	comm := u2fhid.NewCommunication(hiddevice.New(hidDevice), bitboxCMD)
 	device := firmware.NewDevice(nil, nil, &mocks.Config{}, comm, &mocks.Logger{})
 	errpanic(device.Init())
 	device.ChannelHashVerify(true)
